move message types to separate file
